fix: move user filter param to user request schema.
