+++ api/v1/users/users.py
@@ -5,8 +5,8 @@ from pydantic import UUID4
 
 from app.controllers import UserController
 from app.models import UserPermission
-from app.schemas.extra import PaginationResponse, UserFilterParams
-from app.schemas.request import RegisterUserRequest, UpdateUserRequest
+from app.schemas.extra import PaginationResponse
+from app.schemas.request import RegisterUserRequest, UpdateUserRequest, UserFilterParams
 from app.schemas.response import UserResponse
 from core.cache import Cache
 from core.factory import Factory
