+++ app/schemas/extra/__init__.py
@@ -1,5 +1,5 @@
-from .filter import BaseFilterParams, UserFilterParams
+from .filter import BaseFilterParams
 from .pagination import PaginationResponse
 from .token import Token
 
-__all__ = ["Token", "UserFilterParams", "PaginationResponse", "BaseFilterParams"]
+__all__ = ["Token", "PaginationResponse", "BaseFilterParams"]
