+++ app/schemas/request/__init__.py
@@ -1,10 +1,11 @@
 from .auth import UserLoginRequest
 from .post import CreatePostRequest, PostFilterParams, UpdatePostRequest
-from .user import RegisterUserRequest, UpdateUserRequest
+from .user import RegisterUserRequest, UpdateUserRequest, UserFilterParams
 
 __all__ = [
     "RegisterUserRequest",
     "UpdateUserRequest",
+    "UserFilterParams",
     "UserLoginRequest",
     "CreatePostRequest",
     "UpdatePostRequest",
