+++ app/schemas/request/user.py
@@ -1,6 +1,9 @@
+from datetime import datetime
+
 from pydantic import BaseModel, EmailStr, Field
 
 from app.models.user import UserRole
+from app.schemas.extra import BaseFilterParams
 from core.utils import PasswordValidator
 
 
@@ -18,3 +21,15 @@ class UpdateUserRequest(BaseModel):
     first_name: str | None = Field(None, max_length=50, description="Firstname")
     last_name: str | None = Field(None, max_length=50, description="Lastname")
     password: PasswordValidator | None = Field(None, max_length=50, min_length=8, description="Password")
+
+
+class UserFilterParams(BaseFilterParams):
+    email: EmailStr | None = Field(None)
+    role: UserRole | None = Field(None)
+    activated: bool = Field(None)
+    created_from: datetime | None = Field(None)
+    created_to: datetime | None = Field(None)
+    updated_from: datetime | None = Field(None)
+    updated_to: datetime | None = Field(None)
+
+    model_config = {"extra": "forbid"}
